optional append / replace

+++ contextview.py
@@ -26,8 +26,9 @@ BLOCK_HEIGHT = 20
 # Application API
 #######################################################################
 
-def compile_log(log_file, database_file):
-    os.unlink(database_file)
+def compile_log(log_file, database_file, append=False):
+    if not append:
+        os.unlink(database_file)
     db = sqlite3.connect(database_file)
     c = db.cursor()
     c.execute("""
